+++ Assets/Scripts/refactor/EventBusSystem/CombatSystemEvents/CombatSystemApplyDamageToUnitEvent.cs
@@ -1,9 +1,9 @@
 public class CombatSystemApplyDamageToUnitEvent
 {
     public UnitController Target { get; set; }
-    public float Damage { get; set; }
+    public int Damage { get; set; }
 
-    public CombatSystemApplyDamageToUnitEvent(UnitController target, float damage)
+    public CombatSystemApplyDamageToUnitEvent(UnitController target, int damage)
     {
         Target = target;
         Damage = damage;
