+++ Assets/Scripts/refactor/EventBusSystem/EnemyManagerEvents/EnemyController/EnemyControllerAttackEvent.cs
@@ -2,9 +2,9 @@ public class EnemyControllerAttackEvent
 {
     public EnemyController Attacker { get; set; }
     public UnitController Target { get; set; }
-    public float RawDamage { get; set; }
+    public int RawDamage { get; set; }
 
-    public EnemyControllerAttackEvent(EnemyController attacker, UnitController target, float rawDamage)
+    public EnemyControllerAttackEvent(EnemyController attacker, UnitController target, int rawDamage)
     {
         Attacker = attacker;
         Target = target;
