+++ Assets/Scripts/refactor/EventBusSystem/UnitManagerEvents/UnitControllerEvents/UnitControllerAttackEvent.cs
@@ -2,9 +2,9 @@ public class UnitControllerAttackEvent
 {
     public UnitController Attacker { get; set; }
     public EnemyController Target { get; set; }
-    public float RawDamage { get; set; }
+    public int RawDamage { get; set; }
 
-    public UnitControllerAttackEvent(UnitController attacker, EnemyController target, float rawDamage)
+    public UnitControllerAttackEvent(UnitController attacker, EnemyController target, int rawDamage)
     {
         Attacker = attacker;
         Target = target;
